`Cell::partition_id` is stored as `int`

+++ framework/graphs/kba_graph_partitioner.h
@@ -16,9 +16,9 @@ class KBAGraphPartitioner : public GraphPartitioner
 public:
   explicit KBAGraphPartitioner(const InputParameters& params);
 
-  std::vector<int64_t> Partition(const std::vector<std::vector<uint64_t>>& graph,
-                                 const std::vector<Vector3>& centroids,
-                                 int number_of_parts) override;
+  std::vector<int> Partition(const std::vector<std::vector<uint64_t>>& graph,
+                             const std::vector<Vector3>& centroids,
+                             int number_of_parts) override;
 
 protected:
   const size_t nx_, ny_, nz_;
