+++ framework/graphs/linear_graph_partitioner.h
@@ -13,9 +13,9 @@ class LinearGraphPartitioner : public GraphPartitioner
 public:
   explicit LinearGraphPartitioner(const InputParameters& params);
 
-  std::vector<int64_t> Partition(const std::vector<std::vector<uint64_t>>& graph,
-                                 const std::vector<Vector3>& centroids,
-                                 int number_of_parts) override;
+  std::vector<int> Partition(const std::vector<std::vector<uint64_t>>& graph,
+                             const std::vector<Vector3>& centroids,
+                             int number_of_parts) override;
 
 protected:
   const int all_to_rank_;
